requirement.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
langchain-text-splitters
langchain-openai
langchain-huggingface
sentence-transformers
faiss-cpu
chroma
streamlit
langchain_groq
langchain_core
FastAPI
uvicorn
langserve
sse_starlette
bs4